+++ QRCodeEncoderObjectiveCAtGithub/QR_Encode.cpp
@@ -621,7 +621,7 @@ bool CQR_Encode::EncodeData(int nLevel, int nVersion, bool bAutoExtent, int nMas
 	m_nMaskingNo = nMaskingNo;
 
 	// データ長が指定されていない場合は lstrlen によって取得
-	int ncLength = ncSource > 0 ? ncSource : strlen(lpsSource);
+	int ncLength = ncSource > 0 ? ncSource : (int)strlen(lpsSource);
 
 	if (ncLength == 0)
 		return false; // データなし
